@@ -12,48 +12,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.1.2",
-    "@angular/common": "^5.1.2",
-    "@angular/compiler": "^5.1.2",
-    "@angular/core": "^5.1.2",
-    "@angular/forms": "^5.1.2",
-    "@angular/http": "^5.1.2",
-    "@angular/platform-browser": "^5.1.2",
-    "@angular/platform-browser-dynamic": "^5.1.2",
-    "@angular/router": "^5.1.2",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.8",
-    "angularfire2": "^5.0.0-rc.6",
+    "@angular/animations": "^7.2.10",
+    "@angular/common": "^7.2.10",
+    "@angular/compiler": "^7.2.10",
+    "@angular/core": "^7.2.10",
+    "@angular/fire": "^5.1.2",
+    "@angular/forms": "^7.2.10",
+    "@angular/http": "^7.2.10",
+    "@angular/platform-browser": "^7.2.10",
+    "@angular/platform-browser-dynamic": "^7.2.10",
+    "@angular/router": "^7.2.10",
+    "@ng-bootstrap/ng-bootstrap": "^4.1.0",
     "bootstrap": "4.0.0",
     "core-js": "^2.5.3",
-    "firebase": "^4.9.0",
+    "firebase": "^5.9.1",
     "font-awesome": "^4.7.0",
     "jquery": "^3.2.1",
     "lodash": "^4.17.5",
     "popper.js": "^1.12.9",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.4.0",
+    "tslib": "^1.9.0",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "^5.1.2",
-    "@angular/language-service": "^5.1.2",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.6",
+    "@angular/compiler-cli": "^7.2.10",
+    "@angular/language-service": "^7.2.10",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash": "^4.14.106",
     "@types/node": "^9.6.0",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^4.5.0",
     "jasmine": "^3.1.0",
-    "jasmine-core": "~3.1.0",
+    "jasmine-core": "^3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^4.0.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.3.3",
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^1.0.0",
-    "protractor": "~5.3.0",
+    "protractor": "^5.4.2",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "^2.6.2"
+    "typescript": "^3.2.4"
   }
 }
