@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/take';
+import { CanActivate, Router } from '@angular/router';
+import { Observable } from 'rxjs';
+import { map, take } from 'rxjs/operators';
 
 import { AuthService } from '../auth/auth.service';
 
@@ -11,15 +10,16 @@ export class AdminGuard implements CanActivate {
 
   constructor(private authService: AuthService, private router: Router) {}
 
-  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
+  canActivate(): Observable<boolean> | Promise<boolean> | boolean {
     return this.authService.afAuth.authState
-      .take(1)
-      .map((authState) => {
-        if (!authState || !this.authService.isAdmin()) {
-          this.router.navigate(['/']);
-        }
+      .pipe(take(1))
+      .pipe(map((authState) => {
+          if (!authState || !this.authService.isAdmin()) {
+            this.router.navigate(['/']);
+          }
 
-        return !!authState;
-      });
+          return !!authState;
+        })
+      );
   }
 }
