@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import { AngularFireDatabase } from 'angularfire2/database';
-import { Observable } from 'rxjs/Observable';
+import { AngularFireDatabase } from '@angular/fire/database';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
+import { DEADLINE } from '../../constants/global.constants';
 import { AuthService } from '../../auth/auth.service';
 import { BracketChecker } from '../bracket/bracket-checker';
+import { Bracket } from '../../models/bracket.model';
 
 @Component({
   selector: 'app-scoreboard',
@@ -12,25 +15,28 @@ import { BracketChecker } from '../bracket/bracket-checker';
 })
 export class ScoreboardComponent implements OnInit {
   brackets$: Observable<any>;
-  pastDeadline: boolean;
 
-  constructor(private authService: AuthService, private db: AngularFireDatabase) {
-    this.pastDeadline = true;
-  }
+  constructor(private authService: AuthService, private db: AngularFireDatabase) {}
 
   ngOnInit() {
-    this.brackets$ = this.db.list('scoreboard').snapshotChanges().map(changes => {
+    this.brackets$ = this.db.list('scoreboard').snapshotChanges().pipe(map(changes => {
       return changes.map(c => {
-        const bracket = c.payload.val(),
+        const bracket: Partial<Bracket> = c.payload.val(),
           max = 13,
           truncatedName = bracket.name.length > max ? `${bracket.name.substr(0, max)}...` : bracket.name;
 
         return { key: c.payload.key, truncatedName, ...bracket };
       }).sort(BracketChecker.sortBrackets);
-    });
+    }));
   }
 
   isAuthenticated() {
     return this.authService.isAuthenticated();
+  }
+
+  isPastDeadline(): boolean {
+    const today = new Date();
+
+    return today.getTime() > DEADLINE.getTime();
   }
 }
