@@ -1,4 +1,5 @@
 export const KEYS = {
   DUMMY: '-LA-iP4kxzQ65R1x2ad6',
   MASTER: '-LAGXUbTqoKZTg7KeSMn'
-};
+},
+DEADLINE = new Date('04/15/2019 11:59:59 PM');
